#9 230912-1

+++ Jin/Jin/PathMgr.cpp
@@ -11,7 +11,7 @@ void PathMgr::init()
 
 	size_t Len = wcslen(g_szContent);
 
-	for (int i = Len - 1; i >= 0; --i)
+	for (size_t  i = Len - 1; i >= 0; --i)
 	{
 		if ('\\' == g_szContent[i])
 		{
